requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
google-auth==2.23.4
pydantic==2.5.0
pydantic-settings==2.1.0
pymupdf>=1.24.0
pdfminer.six